fix imports so tests don't fail weirdly
